+++ src/Console/Installer.php
@@ -38,18 +38,92 @@ require 'config/paths.php';
  * composer. Customize this class to suit your needs.
  */
 class Installer extends AppInstaller {
-    /**
-     * Does some routine installation tasks so people don't have to
-     * @param \Composer\Script\Event $event The composer event object
-	 * @uses App\Console\Installer::setFolderPermissions()
-	 * @uses createWritableDirectories()
+	/**
+	 * Assets for which create symbolic links.
+	 * The key must be relative to `vendor/`, the value must be relative to `webroot/vendor/`
+	 * @see createSymbolicLinkToVendor()
+	 * @var array
 	 */
-    public static function postInstall(Event $event) {
+	protected static $linksToAssets = [
+		'components/bootstrap-datetimepicker/build'	=> 'bootstrap-datetimepicker',
+		'components/jquery'							=> 'jquery',
+		'components/moment/min'						=> 'moment',
+		'fortawesome/font-awesome'					=> 'font-awesome',
+		'twbs/bootstrap/dist'						=> 'bootstrap'
+	];
+
+	/**
+	 * Creates a symbolic link to vendor asset.
+	 * 
+	 * For example:
+	 * <pre>createSymbolicLinkToVendor('components/jquery', 'jquery');</pre>
+	 * will create a symbolic link from `vendor/components/jquery` to `webroot/vendor/jquery`.
+	 * @param string $from Name, relative to `vendor/`
+	 * @param string $to Name, relative to `webroot/vendor/`
+     * @param \Composer\Script\Event $event The composer event object
+	 */
+	public static function createSymbolicLinkToVendor($from, $to, $event) {
+		$io = $event->getIO();
+		
+		//Creates the target directory (`webroot/vendor/`)
+		if(!file_exists($webrootDir = ROOT.DS.'webroot'.DS.'vendor') && mkdir($webrootDir))
+			$io->write(sprintf('Created `%s` directory', str_replace(ROOT, NULL, $webrootDir)));
+		
+		//Deletes the link, if it already exists
+		if(file_exists($to = $webrootDir.DS.$to))
+			unlink($to);
+		
+		$from = $event->getComposer()->getConfig()->get('vendor-dir').DS.$from;
+				
+		//Creates the symbolic link
+		if(@symlink($from, $to))
+			$io->write(sprintf('Created symbolic link to `%s`', str_replace(ROOT, NULL, $to)));
+
+		else
+			$io->write(sprintf('Failed to create a symbolic link to `%s`', str_replace(ROOT, NULL, $to)));
+	}
+	
+	/**
+	 * Creates some (writable) directories
+     * @param string $dir The application's root directory
+     * @param \Composer\IO\IOInterface $io IO interface to write to console
+	 * @uses App\Console\Installer::createWritableDirectories()
+	 * @uses MeTools\Utility\Thumbs::photo()
+	 * @uses MeTools\Utility\Thumbs::remote()
+	 * @uses MeTools\Utility\Thumbs::video()
+	 */
+    public static function createWritableDirectories($dir, $io) {
+        foreach([
+			WWW_ROOT.'files',
+			dirname(Thumbs::photo()),
+			Thumbs::photo(),
+			Thumbs::remote(),
+			Thumbs::video()
+		] as $path)
+			if(!file_exists($path) && mkdir($path))
+				$io->write(sprintf('Created `%s` directory', str_replace(ROOT, NULL, $path)));
+				
+		//Creates `logs` and `tmp` directories
+		parent::createWritableDirectories($dir, $io);
+    }
+	
+	/**
+	 * Occurs after the autoloader has been dumped, either during install/update, or via the dump-autoload command.
+     * @param \Composer\Script\Event $event The composer event object
+	 * @uses linksToAssets
+	 * @uses createSymbolicLinkToVendor()
+	 * @uses createWritableDirectories()
+	 * @uses setFolderPermissions()
+	 * @see https://getcomposer.org/doc/articles/scripts.md
+	 * @todo Needs to check for root/sudo
+	 */
+	public static function postAutoloadDump(Event $event) {
         $io = $event->getIO();
-
+		
+		//Creates some (writable) directories
         static::createWritableDirectories(ROOT, $io);
-
-        //Asks if the permissions should be changed
+		
+		//If the shell is interactive
         if($io->isInteractive()) {
             $validator = function($arg) {
                 if(in_array($arg, ['Y', 'y', 'N', 'n']))
@@ -57,28 +131,65 @@ class Installer extends AppInstaller {
 				
                 throw new Exception('This is not a valid answer. Please choose Y or n.');
             };
-            $setFolderPermissions = $io->askAndValidate('<info>Set Folder Permissions ? (Default to Y)</info> [<comment>Y,n</comment>]? ', $validator, 10, 'Y');
+			
+			//Asks if the permissions should be changed
+            $setFolderPermissions = $io->askAndValidate('<info>Set folder permissions? (Default to Y)</info> [<comment>Y, n</comment>]? ', $validator, 10, 'Y');
 
             if(in_array($setFolderPermissions, ['Y', 'y']))
-                parent::setFolderPermissions(ROOT, $io);
+                self::setFolderPermissions(ROOT, $io);
+			
+			//Asks if the symbolic links to vendors should be created
+			$createSymbolicLinkToVendor = $io->askAndValidate('<info>Create symbolic links to vendors? (Default to Y)</info> [<comment>Y, n</comment>]? ', $validator, 10, 'Y');
+			
+            if(in_array($createSymbolicLinkToVendor, ['Y', 'y']))
+				foreach(self::$linksToAssets as $from => $to)
+					self::createSymbolicLinkToVendor($from, $to, $event);
         }
-		else
-            parent::setFolderPermissions(ROOT, $io);
-    }
-	
-	/**
-	 * Creates some directories
+		else {
+            self::setFolderPermissions(ROOT, $io);
+
+			foreach(self::$linksToAssets as $from => $to)
+				self::createSymbolicLinkToVendor($from, $to, $event);
+		}
+	}
+
+    /**
+     * Set globally writable permissions on some directories
      * @param string $dir The application's root directory
      * @param \Composer\IO\IOInterface $io IO interface to write to console
-	 * @uses MeTools\Utility\Thumbs::photo()
-	 * @uses MeTools\Utility\Thumbs::remote()
-	 * @uses MeTools\Utility\Thumbs::video()
-	 */
-    public static function createWritableDirectories($dir, $io) {
-        $paths = [dirname(Thumbs::photo()), Thumbs::photo(), Thumbs::remote(), Thumbs::video()];
+     */
+    public static function setFolderPermissions($dir, $io) {
+        //Change the permissions on a path and output the results
+        $changePerms = function ($path, $perms, $io) {
+            //Get current permissions in decimal format so we can bitmask it
+            $currentPerms = octdec(substr(sprintf('%o', fileperms($path)), -4));
+            if(($currentPerms & $perms) == $perms)
+                return;
+			
+            $res = chmod($path, $currentPerms | $perms);
+            if($res)
+                $io->write('Permissions set on '.$path);
+            else
+                $io->write('Failed to set permissions on '.$path);
+        };
 
-        foreach($paths as $path)
-            if(!file_exists($path) && mkdir($path))
-                $io->write(sprintf('Created `%s` directory', str_replace(ROOT, NULL, $path)));
+        $walker = function ($dir, $perms, $io) use (&$walker, $changePerms) {
+            $files = array_diff(scandir($dir), ['.', '..']);
+            foreach ($files as $file) {
+                $path = $dir.'/'.$file;
+
+                if (!is_dir($path))
+                    continue;
+
+                $changePerms($path, $perms, $io);
+                $walker($path, $perms, $io);
+            }
+        };
+
+        $worldWritable = bindec('0000000111');
+        $walker($dir.'/tmp', $worldWritable, $io);
+        $changePerms($dir.'/tmp', $worldWritable, $io);
+        $changePerms($dir.'/logs', $worldWritable, $io);
+        $changePerms($dir.'/webroot/files', $worldWritable, $io);
     }
 }